+++ source/p_portal.cpp
@@ -365,7 +365,7 @@ int P_AddLinkOffset(int startgroup, int targetgroup,
 //
 static bool P_CheckLinkedPortal(portal_t *portal, sector_t *sec)
 {
-   int i = sec - sectors;
+   int i = static_cast<int>(sec - sectors);
 
    if(!portal || !sec)
       return true;
@@ -664,7 +664,7 @@ void P_LinkRejectTable()
       list = groups[i]->seclist;
       for(s = 0; list[s]; s++)
       {
-         int sectorindex1 = list[s] - sectors;
+         int sectorindex1 = static_cast<int>(list[s] - sectors);
 
          for(p = 0; p < groupcount; p++)
          {
@@ -674,7 +674,7 @@ void P_LinkRejectTable()
             list2 = groups[p]->seclist;
             for(q = 0; list2[q]; q++)
             {
-               int sectorindex2 = list2[q] - sectors;
+               int sectorindex2 = static_cast<int>(list2[q] - sectors);
                int pnum = (sectorindex1 * numsectors) + sectorindex2;
 
                rejectmatrix[pnum>>3] &= ~(1 << (pnum&7));
